backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
pillow==10.2.0
python-dotenv==1.0.0

# Google Gemini Vision API
google-generativeai==0.3.2

# Twilio SMS
twilio==8.11.1

# Groq API
groq==0.4.1

# HTTP Requests
httpx==0.26.0
aiofiles==23.2.1

# Database
alembic==1.13.1

# Image Processing
opencv-python-headless==4.9.0.80

# Date/Time
python-dateutil==2.8.2
